+++ src/sit/tu_varna/bg/project/shapes/polygon/Polygon.java
@@ -13,19 +13,19 @@ public class Polygon extends AbstractShape {
 
     /**
      *Конструктор за създаване на многоъгълник
-     * @param fill -> Запълване на многоъгълника
-     * @param strokeWidth ->дебелина на линията описваща многоъгълника
-     * @param stroke ->цвят на линията
-     * @param arrayOfCoordinates -> масив от кординати , редувайки двойката (х,у)
+     * @param fill - Запълване на многоъгълника
+     * @param strokeWidth - дебелина на линията описваща многоъгълника
+     * @param stroke - цвят на линията
+     * @param arrayOfCoordinates - масив от кординати , редувайки двойката (х,у)
      */
-    public Polygon(NamedColors fill, double strokeWidth, NamedColors stroke,List<Integer> arrayOfCoordinates) {
+    public Polygon(NamedColors fill, int strokeWidth, NamedColors stroke,List<Integer> arrayOfCoordinates) {
         super(fill, strokeWidth, stroke);
         this.arrayOfCoordinates = arrayOfCoordinates;
     }
 
     /**
      * Методът конвертира обекта в стрингова репрезентация като запазва полетата на обеката във вид:
-     * <superMethod резултат>  <масив от кориднати>
+     * [superMethod резултат] [масив от кориднати]
      * Методът не използва параметри
      * @return Низ представляваш SVG команда за създаването на многоъгълник
      */
@@ -44,8 +44,8 @@ public class Polygon extends AbstractShape {
         return sb.toString();
     }
     /**
-     * Транслира/Измества фигура по Х кординат
-     * @param sx - > изместване по Х абцисата
+     * Транслира фигура по Х кординат
+     * @param sx -  изместване по Х абцисата
      */
     @Override
     public void translateX(int sx) {
@@ -54,8 +54,8 @@ public class Polygon extends AbstractShape {
         }
     }
     /**
-     * Транслира/Измества фигура по У кординат
-     * @param sy - > изместване по У ординатата
+     * Транслира фигура по У кординат
+     * @param sy - изместване по У ординатата
      */
     @Override
     public void translateY(int sy) {
@@ -65,7 +65,7 @@ public class Polygon extends AbstractShape {
     }
     /**
      * Проверява дали дадената фигура ще е попадне в кръгов регион използвайки формула за кръг:
-     * (x-x0)^2+(y-y0)^2<=R^2
+     * (x-x0)^2+(y-y0)^2=R^2
      * @param region представлява регионът, който методът получава като параметър
      * @return връща дали фигурата е попаднала изцяло в областта
      */
